message_board/requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.2
django-debug-toolbar>=4.2.0
django-ckeditor-5>=0.2.13
pytest>=7.4.3
pytest-django>=4.6.1
gunicorn>=21.2.0
